# Stencil Distribution performance improvments