Reduce the amount of fine-grained communication incurred by `stencilArr.updateFluff` using privitization.

To give an idea of the performance benefits of this PR, the following table shows the runtime improvements on a navier-stokes simulation that uses the stencil distribution with a few different problem sizes:

| problem size | main (sec) | this branch (sec) |
| :--- | :--- | :--- |
| 1024^2 | 9.0031 | 8.23441 |
| 2048^2 | 11.3919 | 8.9551 |
| 4096^2 | 22.3191 | 19.6735 |

I also tested a few other changes to `updateFluff` on the same navier-stokes benchmark. All of these either hurt performance or didn't result in a noticeable improvement:
* using puts instead of gets for the buffer assignment
* using a barrier instead of atomic synchronization
* parallelizing the buffer packing/unpacking
* copying the entire buffers instead of a slice of the first `D` elements

- [x] gasnet paratest